DRY names CRUD controllers
