+++ app/forms/steps/base_names_form.rb
@@ -5,6 +5,12 @@ module Steps
 
     validates_presence_of :new_name, if: :first_name?
 
+    # The default attribute for storing the name in the People table is `full_name`,
+    # but can be customised in subclasses, for example for Children we ask for the 'short' name.
+    def self.name_attribute
+      :full_name
+    end
+
     private
 
     def persist!
@@ -17,12 +23,13 @@ module Steps
     end
 
     def create_new_record
-      record_collection.create(name_attribute => new_name) unless new_name.blank?
+      return if new_name.blank?
+      record_collection.create(self.class.name_attribute => new_name)
     end
 
     def update_existing_record
       names_attributes.each_value do |attrs|
-        next if attrs.fetch(name_attribute.to_s).blank?
+        next if attrs.fetch(self.class.name_attribute.to_s).blank?
         record_collection.update(attrs.fetch('id'), attrs)
       end
     end
@@ -30,12 +37,6 @@ module Steps
     def first_name?
       return if c100_application.nil?
       record_collection.empty?
-    end
-
-    # The default attribute for storing the name in the People table is `full_name`,
-    # but can be customised in subclasses, for example for Children we ask for the 'short' name.
-    def name_attribute
-      :full_name
     end
 
     # :nocov:
